+++ NEWS.md
@@ -8,6 +8,8 @@
 
   Renaming parameters on a per reaction basis is not supported by COPASI.
 
+* Improved consistency of fetching of models from URLs.
+
 # CoRC 0.8.0 (2020-06-25)
 
 * Rebase on COPASI release 4.28 sources.